move test to test_utils.py, install SimpleITK so test runs in coverage
