+++ .github/workflows/code-coverage.yml
@@ -20,6 +20,9 @@ jobs:
       - name: Install coverage
         run: python -m pip install coverage
 
+      - name: Install optional dependencies
+        run: python -m pip install SimpleITK
+
       - name: Unit tests
         run: tests/run_tests.sh -c
